+++ packages/at_commons/lib/at_commons.dart
@@ -32,3 +32,4 @@ export 'package:at_commons/src/enroll/enrollment.dart';
 @experimental
 export 'package:at_commons/src/telemetry/at_telemetry.dart';
 export 'package:at_commons/src/utils/string_utils.dart';
+export 'package:at_commons/src/keystore/public_key_hash.dart';