+++ packages/at_contact/lib/at_contact.dart
@@ -1,6 +1,3 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library at_contact;
 
 export 'package:at_contact/src/at_contacts_impl.dart';