+++ packages/at_onboarding_cli/example/apkam_examples/enroll_app_listen.dart
@@ -65,7 +65,7 @@ Future<void> _notificationCallback(AtNotification notification,
         atAuthKeys.defaultEncryptionPrivateKey!, apkamSymmetricKey);
     var encryptedDefaultSelfEncKey = EncryptionUtil.encryptValue(
         atAuthKeys.defaultSelfEncryptionKey!, apkamSymmetricKey);
-    enrollParamsJson['encryptedDefaultEncryptedPrivateKey'] =
+    enrollParamsJson['encryptedDefaultEncryptionPrivateKey'] =
         encryptedDefaultPrivateEncKey;
     enrollParamsJson['encryptedDefaultSelfEncryptionKey'] =
         encryptedDefaultSelfEncKey;