+++ packages/at_onboarding_cli/lib/src/activate_cli/activate_cli.dart
@@ -1,8 +1,9 @@
+import 'dart:io';
+
+import 'package:args/args.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:at_onboarding_cli/at_onboarding_cli.dart';
-import 'package:args/args.dart';
 import 'package:at_onboarding_cli/src/util/at_onboarding_exceptions.dart';
-import 'dart:io';
 import 'package:at_utils/at_logger.dart';
 
 Future<void> main(List<String> arguments) async {
@@ -10,7 +11,6 @@ Future<void> main(List<String> arguments) async {
   String rootServer = 'root.atsign.org';
   String registrarUrl = 'my.atsign.com';
   AtSignLogger.root_level = 'severe';
-
   //get atSign and CRAM key from args
   final parser = ArgParser()
     ..addOption('atsign', abbr: 'a', help: 'atSign to activate')
@@ -42,8 +42,7 @@ Future<void> main(List<String> arguments) async {
   await activate(argResults);
 }
 
-Future<void> activate(ArgResults argResults,
-    {AtOnboardingService? atOnboardingService}) async {
+Future<void> activate(ArgResults argResults, {AtOnboardingService? atOnboardingService}) async {
   stdout.writeln('[Information] Root server is ${argResults['rootServer']}');
   stdout.writeln(
       '[Information] Registrar url provided is ${argResults['registrarUrl']}');