+++ packages/at_onboarding_cli/test/at_onboarding_cli_test.dart
@@ -29,7 +29,8 @@ void main() {
     test('A test to check authenticate true', () async {
       final atSign = '@alice🛠';
       AtOnboardingPreference onboardingPreference = AtOnboardingPreference()
-        ..atKeysFilePath = 'test/data/@alice🛠.atKeys';
+        ..atKeysFilePath = 'test/data/@alice🛠.atKeys'
+        ..namespace = 'unit_test';
       AtOnboardingService onboardingService =
           AtOnboardingServiceImpl(atSign, onboardingPreference);
       onboardingService.atLookUp = mockAtLookup;